move rotateAuthKey to transactionSubmission
